refactor: api 작업 진행중
