+++ client/src/api/getChatRooms.ts
@@ -1,4 +1,4 @@
-import api from '../../api/customAxios';
+import api from './customAxios';
 
 const getChatRooms = async () => {
   try {
