+++ client/src/api/product-api.ts
@@ -1,0 +1,3 @@
+import api from './customAxios';
+
+export const getUserInformation = () => api.get('/user/me');
