+++ client/src/pages/BorrowWriting.tsx
@@ -11,6 +11,7 @@ import {
   CategoryType,
   descriptionStore,
 } from './../store/PostWriteStore';
+import { getUserInformation } from './../api/product-api';
 
 import HashTagSection from '../components/postWrite/HashTag';
 import ImageUpload from '../components/postWrite/ImageUpload';
@@ -37,19 +38,15 @@ export default function BorrowWriting() {
 
   const navigate = useNavigate();
   // 사용자 가져오기
-  const { data, isLoading } = useQuery(
-    ['userData'],
-    () => api.get('/user/me'),
-    {
-      onError: () => {
-        alert('사용자를 찾을 수 없습니다. \n로그인 화면으로 이동합니다.');
-        navigate('/login');
-      },
-      refetchOnMount: 'always',
-      refetchOnWindowFocus: false,
-      staleTime: 60 * 1000 * 60,
+  const { data, isLoading } = useQuery(['userData'], getUserInformation, {
+    onError: () => {
+      alert('사용자를 찾을 수 없습니다. \n로그인 화면으로 이동합니다.');
+      navigate('/login');
     },
-  );
+    refetchOnMount: 'always',
+    refetchOnWindowFocus: false,
+    staleTime: 60 * 1000 * 60,
+  });
 
   const [filteredCategory, setFilteredCategory] = useState<CategoryType[]>([]);
